pyproject.toml
[tool.poetry]
name = "project1-python"
version = "0.1.0"
description = ""
authors = ["kahpereira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
requests = "^2.31.0"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
